PPSS/Prácticas/P05-Dependencias2/P05-mocks/untitled/src/main/java/ppss/GestorLlamadas.java
package ppss;

public class GestorLlamadas {
    private static final double TARIFA_NOCTURNA=10.5;
    private static final double TARIFA_DIURNA=20.8;


    //De este vamos a crear un partialMock porque pertenece a la misma clase
    public Calendario getCalendario() {
        Calendario c = new Calendario();
        return c;
    }

    public double calculaConsumo(int minutos) {
        Calendario c = getCalendario(); //Dependencia Externa
        int hora = c.getHoraActual(); //Dependencia Externa
        if(hora < 8 || hora > 20) {
            return minutos * TARIFA_NOCTURNA;
        } else {
            return minutos * TARIFA_DIURNA;
        }
    }
}
